Simplify slow/fast nodes using common class
